:recycle: uploadPage 컴포넌트 분리
